requirements.txt
Django~=5.1.1
python-dotenv~=1.0.1
djangorestframework~=3.15.2
djangorestframework-xml~=2.0.0